Rename UnitsTypeEnum to AngleUnitEnum
The enum has been refactored for clarity, since it only concerns angles unit of measure. To maintain the compatibility with configuration files used for gazebo-yarp-plugins this change only affects the name of variables used in the plugin, while leaving the same configuration property names to be used in configuration files.